Add files via upload
In this project, a 3d Design of a coconut climbing machine has done considering the customer requirements-safety, minimizing the climbing time duration, and comfortability. The 3d design has consisted a manual sitting climber with a locking mechanism for the system to the tree safely, ergonomic comfortability of the climber, safe, vertical transfering mechanism.A survey was conducted to identify the customer requirements. My role was to design the body part of the climber machine(Fig: body parts) calculating dimensions for a average human being's sitting position.